pyproject.toml
[tool.poetry]
name = "pdf2epub"
version = "0.1.0"
description = ""
authors = ["morshoto <jojoto8845@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pdfminer-six = "^20240706"
pymupdf = "^1.25.2"
pypdf = "^5.1.0"
beautifulsoup4 = "^4.12.3"
ebooklib = "^0.18"
pillow = "^11.1.0"
colorlog = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
